refactor(): Additional controller for /dashboard/users
