+++ EmployeeWorkOrderManagment/src/main/java/ls/EmployeeWorkOrderManagment/web/controller/DashboardController.java
@@ -19,59 +19,9 @@ import java.util.UUID;
 @RequestMapping("/dashboard")
 public class DashboardController {
 
-    private final UserService userService;
-    private final RoleService roleService;
-
-    public DashboardController(UserService userService, RoleService roleService) {
-        this.userService = userService;
-        this.roleService = roleService;
-    }
-
     @GetMapping
     public String getDashboardPage(Model model) {
         return "dashboard";
     }
 
-    @GetMapping("/users")
-    public String getUserList(Model model,
-                              @ModelAttribute("edit") String editAttribute,
-                              @ModelAttribute("delete") String deleteAttribute) {
-        List<UserSiteRenderDto> users = userService.getAllUsers();
-        model.addAttribute("users", users);
-        Set<RoleDto> roles = roleService.getAllRoles();
-        model.addAttribute("roles", roles);
-        if(!editAttribute.isEmpty()) {
-            model.addAttribute("edit", editAttribute);
-        }
-        if(!deleteAttribute.isEmpty()) {
-            model.addAttribute("delete", deleteAttribute);
-        }
-        return "userList";
-    }
-
-    @PreAuthorize("hasRole('ADMIN')")
-    @GetMapping("/users/delete")
-    public String deleteUserAccount(@RequestParam String id, RedirectAttributes redirectAttributes) {
-        userService.deleteUserAccount(id);
-        redirectAttributes.addAttribute("delete", "User account successfully deleted!");
-        return "redirect:/dashboard/users";
-    }
-
-    @PreAuthorize("hasRole('ADMIN')")
-    @PostMapping("/users")
-    public String editUserAccount(@RequestParam(name = "role") Set<UUID> roles,
-                                  @RequestParam(name = "radio-enabled") boolean enabled,
-                                  @RequestParam(name = "userId") UUID userId,
-                                  Model model,
-                                  RedirectAttributes redirectAttributes) {
-        try{
-            userService.editUserAccount(roles, enabled, userId);
-        } catch (UserNotFoundException exception) {
-            model.addAttribute("message", exception.getMessage());
-            return "message";
-        }
-        redirectAttributes.addAttribute("edit", "Success! User account edited successfully!");
-
-        return "redirect:/dashboard/users";
-    }
 }
